chore: Moved Logo
Was throwing a weird error in Header.tsx but not in CaseStudiesHeader.tsx - I don't know why but resolved now so all good
